requirements.txt
PySide6~=6.6.2
numpy==1.26.4
PySide6-Fluent-Widgets[full]==1.6.6
jinja2~=3.1.4
pyyaml~=6.0.2
jsonschema~=4.23.0
requests~=2.32.3
filetype~=1.2.0
py7zr~=0.22.0
loguru~=0.7.2
attrs~=24.2.0
packaging~=24.1